# HardwareCounterCell::disposable() Performance Improvement