Improve performance of `HardwareCounterCell::disposable()` by 72x.
Since disposable HwCounterCells have to be passed to unmeasured functions, even in hot-paths, performance
is critical.

Dev:
```
Disposable Hw Cell      time:   [170.26 ns 170.48 ns 170.77 ns]
```

This PR:
```
Disposable Hw Cell      time:   [2.3429 ns 2.3600 ns 2.3769 ns]
                        change: [-98.609% -98.601% -98.594%] (p = 0.00 < 0.05)
                        Performance has improved.
```